+++ src/main/java/com/nexmo/client/voice/CallStatus.java
@@ -30,7 +30,11 @@ public enum CallStatus {
     ANSWERED,
     TIMEOUT,
     MACHINE,
-    COMPLETED;
+    COMPLETED,
+    FAILED,
+    REJECTED,
+    BUSY,
+    CANCELLED;
 
     @JsonValue
     @Override